requirements.txt
boto3
elasticsearch
requests
requests-aws4auth
pyyaml
faker
